refactor: rename functions for clarity and consistency; update tests accordingly
