+++ internal/readPolicyFile/main.go
@@ -10,8 +10,8 @@ import (
 	"os"
 )
 
-// --- reads the content of a policy file and returns it as a string ---
-func ReadPolicyFile(filePath string) (string, error) {
+// --- reads the content of a policy file and returns it as a string (with logging) ---
+func ReadPolicyFileWithLogging(filePath string) (string, error) {
 	log.Println("============================================")
 	log.Printf("[INFO] Opening policy file: %s", filePath)
 	file, err := os.Open(filePath)
@@ -45,7 +45,7 @@ func ReadPolicyFile(filePath string) (string, error) {
 }
 
 // --- reads and validates policy file, no logging or side effects ---
-func ReadPolicyFilePure(filePath string) (string, error) {
+func ReadPolicyFile(filePath string) (string, error) {
 	file, err := os.Open(filePath)
 	if err != nil {
 		return "", fmt.Errorf("failed to open policy file: %w", err)
